fix(decky): apply SteamGrid artwork on achievement page
Extract decorateGames into a shared util and reuse it on the
achievement page so the sidebar portrait respects the SteamGrid
setting, mirroring the library behavior.
